Added to CATE.  Took out TOSC.

+++ tools/CATE.ps1
@@ -1,7 +1,7 @@
 
 <#PSScriptInfo
 
-.VERSION 5.7
+.VERSION 5.9
 
 .GUID f842f577-3f42-4cb0-91e7-97b499260a21
 
@@ -527,33 +527,68 @@ CATE-Delete-Files-Only ($envSystemRoot + '\inf') '*.log'
 
 CATE-Delete-Files-Only ($envSystemRoot + '\Prefetch') '*.pf'
 
-# Compact the Windows Search database, if found
+function Background-Run-For-Five-Minutes-Max {
+	param( $MyScript )
+	
+	$JobGUID = [system.Guid]::NewGuid()
 
-$wsdb = "$env:AllUsersProfile" + "\Microsoft\Search\Data\Applications\Windows\Windows.edb"
-if (Test-Path -Path $wsdb -PathType Leaf) {
-	Write-Output "Compacting the Windows Search database..."
+	$elapsedEventHandler = {
+		param ([System.Object]$sender, [System.Timers.ElapsedEventArgs]$e)
 
-	Write-Output "Stopping service..."
-	Stop-Service wsearch -ErrorAction SilentlyContinue *> $null
-	Write-Output "Compacting..."
-	& esentutl.exe /d $wsdb
-	Write-Output "Setting service for delayed start..."
-	start-process -FilePath sc.exe -ArgumentList 'config wsearch start=delayed-auto' -ErrorAction SilentlyContinue *> $null
-	Write-Output "Restarting service..."
-	Start-Service wsearch -ErrorAction SilentlyContinue *> $null
+		($sender -as [System.Timers.Timer]).Stop()
+		Unregister-Event -SourceIdentifier $JobGUID
+		Write-Host "Job $JobGUID removed by force as it exceeded timeout!"
+		Get-Job -Name $JobGUID | Remove-Job -Force
+	}
+
+	$timer = New-Object System.Timers.Timer -ArgumentList 300000 #just change the timeout here
+	Register-ObjectEvent -InputObject $timer -EventName Elapsed -Action $elapsedEventHandler -SourceIdentifier $JobGUID
+	$timer.Start()
+
+	Start-Job -ScriptBlock $MyScript -Name $JobGUID
 }
 
-Write-Output "Compacting the Windows Update database..."
+# Compact the Windows Search database, if found
 
-Write-Output "Stopping services..."
-Stop-Service wuauserv *> $null
-Stop-Service bits *> $null
-Write-Output "Compacting..."
-& esentutl.exe /d ("$env:WINDIR" + '\SoftwareDistribution\DataStore\DataStore.edb')
-Write-Output "Restarting services..."
-Start-Service wuauserv *> $null
-Start-Service bits *> $null
+Write-Output "Setting up background job to compact the Windows Search database, will run 5 minutes max..."
 
+$CompactWindowsSearchDatabase = {
+	$wsdb = "$env:AllUsersProfile" + "\Microsoft\Search\Data\Applications\Windows\Windows.edb"
+	if (Test-Path -Path $wsdb -PathType Leaf) {
+		Write-Output "Stopping service..."
+		Stop-Service wsearch -ErrorAction SilentlyContinue *> $null
+		Write-Output "Compacting..."
+		& esentutl.exe /d $wsdb
+		Write-Output "Setting service for delayed start..."
+		start-process -FilePath sc.exe -ArgumentList 'config wsearch start=delayed-auto' -ErrorAction SilentlyContinue *> $null
+		Write-Output "Restarting service..."
+		Start-Service wsearch -ErrorAction SilentlyContinue *> $null
+	}
+}
+	
+Background-Run-For-Five-Minutes-Max $CompactWindowsSearchDatabase
+
+# Compact the Windows Update database, if found
+
+Write-Output ''
+Write-Output "Setting up background job to compact the Windows Update database, will run 5 minutes max..."
+Write-Output ''
+
+$CompactWindowsUpdateDatabase = {
+	$wudb = "$env:AllUsersProfile" + '\SoftwareDistribution\DataStore\DataStore.edb'
+	if (Test-Path -Path $wudb -PathType Leaf) {
+		Write-Output "Stopping services..."
+		Stop-Service wuauserv *> $null
+		Stop-Service bits *> $null
+		Write-Output "Compacting..."
+		& esentutl.exe /d $wudb
+		Write-Output "Restarting services..."
+		Start-Service wuauserv *> $null
+		Start-Service bits *> $null
+	}
+}
+	
+Background-Run-For-Five-Minutes-Max $CompactWindowsUpdateDatabase
 
 ""
 ""
@@ -641,107 +676,3 @@ exit
 
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-# SIG # Begin signature block
-# MIIQIQYJKoZIhvcNAQcCoIIQEjCCEA4CAQExCzAJBgUrDgMCGgUAMGkGCisGAQQB
-# gjcCAQSgWzBZMDQGCisGAQQBgjcCAR4wJgIDAQAABBAfzDtgWUsITrck0sYpfvNR
-# AgEAAgEAAgEAAgEAAgEAMCEwCQYFKw4DAhoFAAQU/Nr/RnjOdi2uudzAeAjYr1bg
-# 1VqgggupMIIDDDCCAfSgAwIBAgIQGsDSayfiep9MU0Gn9qSESTANBgkqhkiG9w0B
-# AQsFADAeMRwwGgYDVQQDDBNDQlQgUG93ZXJTaGVsbCBDb2RlMB4XDTIwMDgyNjE0
-# Mjc1MloXDTI1MDgyNjE0Mzc1MlowHjEcMBoGA1UEAwwTQ0JUIFBvd2VyU2hlbGwg
-# Q29kZTCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAN2yXYLb9UAxirdD
-# q1wGZNWYCUGXlju/a3aZmN8iVpJRAxr1IsJmeEjgZgvhKxCFiQOulI1XzfC0qF3E
-# W5wSZbmoXlHkeihS3Qne4735WEAAp0JVVsdstZuK3faFJCtaWorhtHF9QTfPmMZb
-# CU40CNmZmaH/FaAvwX7Y0lt2bNLbf6ICgwwMQy1KOQIQtgIZYohBM9ceC+OyU/ko
-# VVvMFQeX9V+vYijAWJxlpTv+A52Z4MvlkpO+zdVKBm3pg0BVxK/jgk288K9otTCY
-# AybawIjeVS77C//wGwQjM+Qgpdswhxnu9AYUtREht9aaqL2trA3eIQhdzFCMpTze
-# L3KEFqECAwEAAaNGMEQwDgYDVR0PAQH/BAQDAgeAMBMGA1UdJQQMMAoGCCsGAQUF
-# BwMDMB0GA1UdDgQWBBTuCNhmsoGL4xfXdeHB9ajub1TGKjANBgkqhkiG9w0BAQsF
-# AAOCAQEATsu+0CzTZUEniNgB+6gJ5Ver2WexHfVumGqXQ/VgSV6o7Na1adM5YDEl
-# OOupwxAqYyHRh7S6ljOTqSxs0MXv0u3N9WuL8QKTZi6N8b6Fm9721TZiszxDdefW
-# ZxgSkRq9UAsx9yyVEq6JfKenH/JmPPZ/a7E0ioYTkAHNR3CHTh8WeeQBBWpFDO7X
-# mtAjrSbNTymvE+zAQ8t6BuxAua5JzGrBGbEKaTMhwFP/nseeCMKRNsH6EevnYah/
-# +DmW6AzXJOuoazW+KNF54T6drKTx1lLBFfTjNSCKwN6Z82jpFfCD1/XDWC8njt/N
-# 766cq+efmqOajiWX6uQSIIPH3Oy9cDCCA+4wggNXoAMCAQICEH6T6/t8xk5Z6kua
-# d9QG/DswDQYJKoZIhvcNAQEFBQAwgYsxCzAJBgNVBAYTAlpBMRUwEwYDVQQIEwxX
-# ZXN0ZXJuIENhcGUxFDASBgNVBAcTC0R1cmJhbnZpbGxlMQ8wDQYDVQQKEwZUaGF3
-# dGUxHTAbBgNVBAsTFFRoYXd0ZSBDZXJ0aWZpY2F0aW9uMR8wHQYDVQQDExZUaGF3
-# dGUgVGltZXN0YW1waW5nIENBMB4XDTEyMTIyMTAwMDAwMFoXDTIwMTIzMDIzNTk1
-# OVowXjELMAkGA1UEBhMCVVMxHTAbBgNVBAoTFFN5bWFudGVjIENvcnBvcmF0aW9u
-# MTAwLgYDVQQDEydTeW1hbnRlYyBUaW1lIFN0YW1waW5nIFNlcnZpY2VzIENBIC0g
-# RzIwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQCxrLNJVEuXHBIK2CV5
-# kSJXKm/cuCbEQ3Nrwr8uUFr7FMJ2jkMBJUO0oeJF9Oi3e8N0zCLXtJQAAvdN7b+0
-# t0Qka81fRTvRRM5DEnMXgotptCvLmR6schsmTXEfsTHd+1FhAlOmqvVJLAV4RaUv
-# ic7nmef+jOJXPz3GktxK+Hsz5HkK+/B1iEGc/8UDUZmq12yfk2mHZSmDhcJgFMTI
-# yTsU2sCB8B8NdN6SIqvK9/t0fCfm90obf6fDni2uiuqm5qonFn1h95hxEbziUKFL
-# 5V365Q6nLJ+qZSDT2JboyHylTkhE/xniRAeSC9dohIBdanhkRc1gRn5UwRN8xXnx
-# ycFxAgMBAAGjgfowgfcwHQYDVR0OBBYEFF+a9W5czMx0mtTdfe8/2+xMgC7dMDIG
-# CCsGAQUFBwEBBCYwJDAiBggrBgEFBQcwAYYWaHR0cDovL29jc3AudGhhd3RlLmNv
-# bTASBgNVHRMBAf8ECDAGAQH/AgEAMD8GA1UdHwQ4MDYwNKAyoDCGLmh0dHA6Ly9j
-# cmwudGhhd3RlLmNvbS9UaGF3dGVUaW1lc3RhbXBpbmdDQS5jcmwwEwYDVR0lBAww
-# CgYIKwYBBQUHAwgwDgYDVR0PAQH/BAQDAgEGMCgGA1UdEQQhMB+kHTAbMRkwFwYD
-# VQQDExBUaW1lU3RhbXAtMjA0OC0xMA0GCSqGSIb3DQEBBQUAA4GBAAMJm495739Z
-# MKrvaLX64wkdu0+CBl03X6ZSnxaN6hySCURu9W3rWHww6PlpjSNzCxJvR6muORH4
-# KrGbsBrDjutZlgCtzgxNstAxpghcKnr84nodV0yoZRjpeUBiJZZux8c3aoMhCI5B
-# 6t3ZVz8dd0mHKhYGXqY4aiISo1EZg362MIIEozCCA4ugAwIBAgIQDs/0OMj+vzVu
-# BNhqmBsaUDANBgkqhkiG9w0BAQUFADBeMQswCQYDVQQGEwJVUzEdMBsGA1UEChMU
-# U3ltYW50ZWMgQ29ycG9yYXRpb24xMDAuBgNVBAMTJ1N5bWFudGVjIFRpbWUgU3Rh
-# bXBpbmcgU2VydmljZXMgQ0EgLSBHMjAeFw0xMjEwMTgwMDAwMDBaFw0yMDEyMjky
-# MzU5NTlaMGIxCzAJBgNVBAYTAlVTMR0wGwYDVQQKExRTeW1hbnRlYyBDb3Jwb3Jh
-# dGlvbjE0MDIGA1UEAxMrU3ltYW50ZWMgVGltZSBTdGFtcGluZyBTZXJ2aWNlcyBT
-# aWduZXIgLSBHNDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAKJjCzlE
-# uLsjp0RJuw7/ofBhClOTsJjbrSwPSsVu/4Y8U1UPFc4EPyv9qZaW2b5heQtbyUyG
-# duXgQ0sile7CK0PBn9hotI5AT+6FOLkRxSPyZFjwFTJvTlehroikAtcqHs1L4d1j
-# 1ReJMluwXplaqJ0oUA4X7pbbYTtFUR3PElYLkkf8q672Zj1HrHBy55LnX80QucSD
-# ZJQZvSWA4ejSIqXQugJ6oXeTW2XD7hd0vEGGKtwITIySjJEtnndEH2jWqHR32w5b
-# MotWizO92WPISZ06xcXqMwvS8aMb9Iu+2bNXizveBKd6IrIkri7HcMW+ToMmCPsL
-# valPmQjhEChyqs0CAwEAAaOCAVcwggFTMAwGA1UdEwEB/wQCMAAwFgYDVR0lAQH/
-# BAwwCgYIKwYBBQUHAwgwDgYDVR0PAQH/BAQDAgeAMHMGCCsGAQUFBwEBBGcwZTAq
-# BggrBgEFBQcwAYYeaHR0cDovL3RzLW9jc3Aud3Muc3ltYW50ZWMuY29tMDcGCCsG
-# AQUFBzAChitodHRwOi8vdHMtYWlhLndzLnN5bWFudGVjLmNvbS90c3MtY2EtZzIu
-# Y2VyMDwGA1UdHwQ1MDMwMaAvoC2GK2h0dHA6Ly90cy1jcmwud3Muc3ltYW50ZWMu
-# Y29tL3Rzcy1jYS1nMi5jcmwwKAYDVR0RBCEwH6QdMBsxGTAXBgNVBAMTEFRpbWVT
-# dGFtcC0yMDQ4LTIwHQYDVR0OBBYEFEbGaaMOShQe1UzaUmMXP142vA3mMB8GA1Ud
-# IwQYMBaAFF+a9W5czMx0mtTdfe8/2+xMgC7dMA0GCSqGSIb3DQEBBQUAA4IBAQB4
-# O7SRKgBM8I9iMDd4o4QnB28Yst4l3KDUlAOqhk4ln5pAAxzdzuN5yyFoBtq2MrRt
-# v/QsJmMz5ElkbQ3mw2cO9wWkNWx8iRbG6bLfsundIMZxD82VdNy2XN69Nx9DeOZ4
-# tc0oBCCjqvFLxIgpkQ6A0RH83Vx2bk9eDkVGQW4NsOo4mrE62glxEPwcebSAe6xp
-# 9P2ctgwWK/F/Wwk9m1viFsoTgW0ALjgNqCmPLOGy9FqpAa8VnCwvSRvbIrvD/niU
-# UcOGsYKIXfA9tFGheTMrLnu53CAJE3Hrahlbz+ilMFcsiUk/uc9/yb8+ImhjU5q9
-# aXSsxR08f5Lgw7wc2AR1MYID4jCCA94CAQEwMjAeMRwwGgYDVQQDDBNDQlQgUG93
-# ZXJTaGVsbCBDb2RlAhAawNJrJ+J6n0xTQaf2pIRJMAkGBSsOAwIaBQCgeDAYBgor
-# BgEEAYI3AgEMMQowCKACgAChAoAAMBkGCSqGSIb3DQEJAzEMBgorBgEEAYI3AgEE
-# MBwGCisGAQQBgjcCAQsxDjAMBgorBgEEAYI3AgEVMCMGCSqGSIb3DQEJBDEWBBQX
-# IWDd1KdzBqy8KN4/YBjHRZfVvDANBgkqhkiG9w0BAQEFAASCAQDBTdblTRtef6sp
-# W3EZJ1DJ1QeWHz7tgowD1UnpCNvMRYz8V7pXb9vpw3WbuvcWrEqHCO6ceUc2ox5X
-# nz7o5mMcsnYQDFqHZuythTrpXQGFX4xNsQNB6bdzJT5M1AIjQfa7BL+l/qGTA622
-# bU0VgWFdL0EW4KxZlpHNAJfj+aaWji9Yxpo8p3MXf78rFGjnoU6Fl6KeuB6p0aT+
-# W8EBiZVejTXSI48Pxkp3sKr1JwVtoOX1JemugXcw0Q3Oln+HdixTeNvSEUPNjCBA
-# C/JKJr+MdfbrBc9XjgqRt01c9xbY85aDa4p+QxgEHHkDk05wB9S8Oe0Z8tZnC81j
-# cvku8tYBoYICCzCCAgcGCSqGSIb3DQEJBjGCAfgwggH0AgEBMHIwXjELMAkGA1UE
-# BhMCVVMxHTAbBgNVBAoTFFN5bWFudGVjIENvcnBvcmF0aW9uMTAwLgYDVQQDEydT
-# eW1hbnRlYyBUaW1lIFN0YW1waW5nIFNlcnZpY2VzIENBIC0gRzICEA7P9DjI/r81
-# bgTYapgbGlAwCQYFKw4DAhoFAKBdMBgGCSqGSIb3DQEJAzELBgkqhkiG9w0BBwEw
-# HAYJKoZIhvcNAQkFMQ8XDTIwMDgyNjE0Mzc1NlowIwYJKoZIhvcNAQkEMRYEFGU/
-# BmgE+/c9F4OjC17wHXBftwNkMA0GCSqGSIb3DQEBAQUABIIBAFxqLIv4/gOd7gDc
-# lXtQhIWd+TmhaL98K0grCkv5e+RFKHZmUgX5XNL5koI4EQVbKBvZ8Bhx+Go2CT59
-# KTy6GvekCL8I2uQfdyBpiTIPbqWf0uPa5Ss0Bt+g5g9jFU1Rgeebah/PiMkSYe8N
-# KQ6ouhQ/sNOBratWZ6Tb7nEMKAcZLsaKpdIMPJLp3UEI9YiTUGHeS8hi+HxFWRLK
-# ifPdlThhj9J33OxwPs2Q/+VXzzS1v9l0EsKoPVCibYme+6eLM0MBWfEh2u7mpfoe
-# HD0szdVsMg8nYN6VWIRZ29+JRm/nqRmkWf1yj2nS8f9mP2e70GIB4RmnYD+PNOrb
-# iPdfVTk=
-# SIG # End signature block
